Move static exchange market config, market ids, and market names to testutil/daemons/pricefeed/exchange_config.
