--- protocol/daemons/pricefeed/client/constants/exchange_common/exchange_names.go
@@ -1,20 +1,0 @@
-package exchange_common
-
-// Names of each exchange - used for logging and errors.
-const (
-	EXCHANGE_NAME_BINANCE                = "Binance"
-	EXCHANGE_NAME_BINANCEUS              = "BinanceUS"
-	EXCHANGE_NAME_BITFINEX               = "Bitfinex"
-	EXCHANGE_NAME_KRAKEN                 = "Kraken"
-	EXCHANGE_NAME_GATE                   = "Gate"
-	EXCHANGE_NAME_BITSTAMP               = "Bitstamp"
-	EXCHANGE_NAME_BYBIT                  = "Bybit"
-	EXCHANGE_NAME_CRYPTO_COM             = "CryptoCom"
-	EXCHANGE_NAME_HUOBI                  = "Huobi"
-	EXCHANGE_NAME_KUCOIN                 = "Kucoin"
-	EXCHANGE_NAME_OKX                    = "Okx"
-	EXCHANGE_NAME_MEXC                   = "Mexc"
-	EXCHANGE_NAME_COINBASE_PRO           = "CoinbasePro"
-	EXCHANGE_NAME_TEST_EXCHANGE          = "TestExchange"
-	EXCHANGE_NAME_TEST_VOLATILE_EXCHANGE = "TestVolatileExchange"
-)
