+++ protocol/daemons/pricefeed/client/price_function/coinbase_pro/coinbase_pro.go
@@ -48,7 +48,7 @@ func CoinbaseProPriceFunction(
 	// Get ticker. The API response should only contain information for one market.
 	ticker, _, err := price_function.GetOnlyTickerAndExponent(
 		tickerToExponent,
-		exchange_common.EXCHANGE_NAME_COINBASE_PRO,
+		exchange_common.EXCHANGE_ID_COINBASE_PRO,
 	)
 	if err != nil {
 		return nil, nil, err
