+++ protocol/testutil/daemons/pricefeed/exchange_config/market_id.go
@@ -1,4 +1,4 @@
-package exchange_common
+package exchange_config
 
 import "github.com/dydxprotocol/v4-chain/protocol/daemons/pricefeed/client/types"
